Merged the top-level tts.py into src/chatterbox/tts.py.
Also made a bunch of small cleanups and very minor optimizations.

When using this as a library, I was originally thinking that the top-level tts.py would be incorporated directly into the caller. But that doesn't really add any value; it's easier to simply "from chatterbox.tts import TTS".

This will also make it much easier to further customize the library.