Removed the dependency on librosa.
Librosa is bloated, and pulls in other dependencies such as Randa. That currently doesn't support NumPy 2.3, even though many other libraries require it, creating conflicts.

This replaces all uses of librosa with equivalent torchaudio calls.

That required a bunch of related clean-ups to operate on tensors end-to-end rather than constantly converting from NumPy arrays to tensors and vice-versa.

There are more such opportunities, especially in how s3gen needlessly resamples the reference audio that we've already previously resampled. That will happen in later commits; this one is large enough already.

The only librosa call that wasn't replaced with a perfect equivalent was librosa.load, to load the reference audio. It was replaced using Python's native wave library.

That limits reference audio to uncompressed PCM wav files, which is fine; that's the best format anyway. I could have used torchaudio.load, but it will soon be deprecated in favor of torchcodecs, and I'm hoping to avoid adding that new dependency later.